requirements.txt
# FastAPI
fastapi
uvicorn

# Database
sqlalchemy
databases
asyncpg

# Authentication
python-jose
passlib

# Cryptography
cryptography

# Others
pydantic

# Blockchain and IPFS
web3
requests
requests-toolbelt
python-dotenv